pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cortex-knowledge-assistant"
version = "0.1.0-beta"
description = "Enterprise-grade RAG system with streaming LLM responses, DLP/PII protection, and role-based access control"
readme = "README.md"
license = {text = "AGPL-3.0-or-later"}
authors = [
  {name = "DeepRatAI", email = "deeprat.tec@gmail.com"}
]
maintainers = [
  {name = "DeepRatAI", email = "deeprat.tec@gmail.com"}
]
keywords = ["rag", "llm", "ai", "knowledge-assistant", "fastapi", "enterprise", "dlp", "pii", "security"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Web Environment",
  "Framework :: FastAPI",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Typing :: Typed"
]
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.1",
  "langchain>=0.2.0",
  "langchain-community>=0.2.0",
  "langchain-ollama>=0.1.0",
  "qdrant-client>=1.9.0,<1.13.0",
  "redis>=5.0.0",
  "httpx>=0.27.0",
  "orjson>=3.9.15",
  "structlog>=24.1.0",
  "sentence-transformers>=2.5.1",
  "prometheus-client>=0.20.0"
  ,"opentelemetry-api>=1.25.0"
  ,"opentelemetry-sdk>=1.25.0"
  ,"opentelemetry-instrumentation-fastapi>=0.50b0"
  ,"opentelemetry-exporter-otlp>=1.25.0"
  ,"SQLAlchemy>=2.0.44"
  ,"passlib[bcrypt]>=1.7.4"
  ,"bcrypt>=4.0.0,<4.1.0"
  ,"PyJWT>=2.10.1"
  ,"pypdf>=4.0.0"
  ,"python-multipart>=0.0.6"
  ,"psycopg[binary]>=3.1.0"
  ,"APScheduler>=3.10.0,<4.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "flake8>=7.0.0",
  "pylint>=3.2.0",
  "black>=24.4.0",
  "isort>=5.13.0",
  "mypy>=1.10.0",
  "pip-audit>=2.7.0",
  "types-redis",
  "types-requests"
]
perf = [
  "tiktoken>=0.8.0"
]

[tool.black]
line-length = 120
target-version = ["py311", "py312"]

[tool.isort]
profile = "black"
line_length = 120
known_first_party = ["src"]

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]
max-complexity = 20
exclude = [".git", "__pycache__", ".venv", "build", "dist"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_return_any = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "e2e: marks tests as end-to-end tests",
  "integration: marks tests as integration tests"
]

[project.urls]
Homepage = "https://github.com/DeepRatAI/cortex-knowledge-assistant"
Documentation = "https://github.com/DeepRatAI/cortex-knowledge-assistant#readme"
Repository = "https://github.com/DeepRatAI/cortex-knowledge-assistant.git"
Issues = "https://github.com/DeepRatAI/cortex-knowledge-assistant/issues"
Changelog = "https://github.com/DeepRatAI/cortex-knowledge-assistant/blob/main/CHANGELOG.md"
